backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
firebase-admin==6.5.0
google-cloud-firestore==2.16.1
requests==2.32.3
PyJWT==2.9.0
openai

